-Écriture de la quasi totalité de la documentation -Réorganisation de la structure -Création d'objets permettant de créer des scripts et des actions parallèles au jeu -Nettoyage important-Écriture de la quasi totalité de la documentation -Création d'objets permettant de créer des scripts et des actions parallèles au jeu -Nettoyage important -Optimisation des boutons afin d'améliorer les performances et la propreté -Ajout de fonctionnalités liées au trajectoires -Amélioration des réactions aux collisions -Ajout d'une interface de développement activable grâce à la touche "supprimer" du clavier -Ajout de méthode permettant de supprimer des entités du jeu et de l'affichage -Limitation de la fréquence de rafraîchissement de l'écran à 120Hz -Optimisation des boutons afin d'améliorer les performances et la propreté -Ajout de fonctionnalités liées au trajectoires -Amélioration des réactions aux collisions -Ajout d'une interface de développement activable grâce à la touche "supprimer" du clavier -Ajoute de méthode permettant de supprimer des entités du jeu et de l'affichage -Augmentation de la précision des déplacements -Ajout d'un Objet représentant une particule -Amélioration des performances et de la stabilité des sons
-Notes par rapport aux performances:
Sans la limite de fréquence d'affichage, l'affichage se fait à 200-250 ips.
Avec la limite, l'affichage se fait à 120 ips, cela permet aux calculs logiques d'être plus rapides
Les sons ne bloquent plus le jeu lors de leur arrêt.

+++ src/edu/bouyaka/engine/interfaces/Movings.java
@@ -18,21 +18,21 @@ public interface Movings {
 
 	public int getRightKey();
 
-	public void moveUp(int n);
+	public void moveUp(double n);
 
-	public void moveDown(int n);
+	public void moveDown(double n);
 
-	public void moveLeft(int n);
+	public void moveLeft(double n);
 
-	public void moveRight(int n);
+	public void moveRight(double n);
 
-	public void moveUpRight(int n);
+	public void moveUpRight(double n);
 
-	public void moveDownRight(int n);
+	public void moveDownRight(double n);
 
-	public void moveUpLeft(int n);
+	public void moveUpLeft(double n);
 
-	public void moveDownLeft(int n);
+	public void moveDownLeft(double n);
 
 	public void move();
